Fixes certain cases where the Lambdas array was the incorrect size.
In certain cases the Lambdas array was not kept in sync with the
size of the BasePointers array, leading to a mismatch between array
indexes when generating offload information for lambda closures.

This fixes the issue by updating the size of the Lambdas array to
be the same size as the BasePointers array (with nullptr added in
the extra locations) after a specific function call that modifies
the size of the BasePointers array.