Port argument passing to device toolchains
This is based on https://reviews.llvm.org/D34784 and includes
recent cleanups and fixes from upstream.

The patch also allows to drop the hardcoded GPU architecture selection.
This can now be done with command line arguments like:
-fopenmp-targets=nvptx64-nvidia-cuda -Xopenmp-target -march=sm_60

In addition, the default architecture can be set with the CMake
variable CLANG_OPENMP_NVPTX_DEFAULT_ARCH that defaults to sm_30.