Emit omp.init.ds after all private variables for the teams directive
We had to do some crazy stuff to find the right spot when at
the beginning of the distribute directive.

This change builds working binaries when using reductions on array
sections in an SPMD kernel, see added test case. For generic kernels,
it currently leads to an infinite loop when moving instructions...