Stop moving PHINodes for data sharing
This led to an infinite loop when using reductions on array sections
in generic kernels because the initialization is done in a loop with
a PHINode.
After this change, the binary is built successfully but produces
incorrect results because we currently cannot handle multiple
privatizations, see https://puna0.watson.ibm.com/T294.